Azure/vj_azure_requirements.txt
azure-mgmt-resource==21.1.0
azure-mgmt-compute==28.0.0
azure-mgmt-network==21.0.1
azure-mgmt-storage==20.1.0
azure-identity==1.11.0
